# [SPARK-32025][SQL] Csv schema inference problems with different types in the same column