--- a/sql/core/src/test/resources/test-data/mixed-types1.csv
+++ b/sql/core/src/test/resources/test-data/mixed-types1.csv
@@ -1,4 +1,0 @@
-col_mixed_types
-2012
-1997
-True