@planga82 2020-06-22
@MaxGekk Yes, the last type in inference is StringType in all inference chains. I have change the description to show it better.
The principal difference is than now we have different chains instead of one