@cloud-fan 2020-06-23
Does JSON source have the same problem?