@planga82 2020-06-23
@HyukjinKwon It's exactly what you say, it only happens when the incompatibility is inside one partition. I will change the PR to use compatibleType, and I will make some performance test. Thanks for your help!

@cloud-fan I tested the same situation with json and it works fine, we don't have problems there