@planga82 2020-06-23
I have done some performance tests in my local machine
- File: 999999999 lines integers / 10 GB 
- Resources: 4 Cores

Result without changes:
- 21m10,114s
- 20m18,626s
- 19m46,158s

Results after changes
- 19m46,388s
- 22m10,784s
- 21m15,937s

It seems that we don't have a very significant impact but the tests in local are not the best way to be sure